Support returning to title.
Removes all circular references I could figure out, mostly of the parent-child relationship kind. For things like the config and Events object, use a static singleton pattern for MainClass. Decided to get rid of the events that MainClass was registering on the ObjectLists to prevent that circular ref.

Furthermore assist with garbage collection by explicitly Diposing certain large object graphs, namely the ObjectLists and the WorldGraph.